+++ cvar.c
@@ -35,9 +35,10 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 
 static void Cvar_ApplyLatchedUpdate(cvar_t* var);
 
-extern void CL_UserinfoChanged (char *key, char *value);
-extern void SV_ServerinfoChanged (char *key, char *value);
-extern void Help_DescribeCvar (cvar_t *v);
+void CL_UserinfoChanged(char *key, char *value);
+void SV_ServerinfoChanged(char *key, char *value);
+void Help_DescribeCvar(cvar_t *v);
+void VID_ReloadCvarChanged(cvar_t* var);
 
 extern cvar_t r_fullbrightSkins;
 extern cvar_t cl_fakeshaft;
@@ -258,6 +259,9 @@ void Cvar_SetEx(cvar_t *var, char *value, qbool ignore_callback)
 	if (var->flags & CVAR_RECOMPILE_PROGS) {
 		renderer.CvarForceRecompile(var);
 	}
+	if ((var->flags & CVAR_RELOAD_GFX) && host_everything_loaded) {
+		VID_ReloadCvarChanged(var);
+	}
 #endif
 }
 
@@ -377,8 +381,9 @@ void Cvar_Register(cvar_t *var)
 		}
 
 		// warn if CVAR_SILENT is not set
-		if (!(old->flags & CVAR_SILENT))
+		if (!(old->flags & CVAR_SILENT)) {
 			Com_Printf("Can't register variable %s, already defined\n", var->name);
+		}
 
 		return;
 	}
@@ -1618,6 +1623,7 @@ void Cvar_ExecuteQueuedChanges(void)
 {
 	cvar_t* cvar;
 	qbool vid_restart = false;
+	qbool vid_reload = false;
 	qbool sound_restart = false;
 
 	for (cvar = cvar_vars; cvar; cvar = cvar->next) {
@@ -1635,13 +1641,44 @@ void Cvar_ExecuteQueuedChanges(void)
 		}
 
 		vid_restart |= (cvar->flags & CVAR_LATCH_GFX) && cvar->latchedString;
+		vid_reload |= (cvar->flags & CVAR_RELOAD_GFX && cvar->modified);
 		sound_restart |= (cvar->flags & CVAR_LATCH_SOUND) && cvar->latchedString;
 	}
 
 	if (vid_restart) {
 		Cbuf_AddTextEx(&cbuf_main, "\nvid_restart\n");
 	}
-	else if (sound_restart) {
-		Cbuf_AddTextEx(&cbuf_main, "\ns_restart\n");
+	else {
+		if (sound_restart) {
+			Cbuf_AddTextEx(&cbuf_main, "\ns_restart\n");
+		}
+		if (vid_reload) {
+			Cbuf_AddTextEx(&cbuf_main, "\nvid_reload\n");
+		}
 	}
+
+	Cvar_ClearAllModifiedFlags(CVAR_RELOAD_GFX);
+}
+
+void Cvar_ClearAllModifiedFlags(int flags)
+{
+	cvar_t* cvar;
+
+	for (cvar = cvar_vars; cvar; cvar = cvar->next) {
+		if (cvar->flags & flags) {
+			cvar->modified = false;
+		}
+	}
+}
+
+qbool Cvar_AnyModified(int flags)
+{
+	cvar_t* cvar;
+
+	for (cvar = cvar_vars; cvar; cvar = cvar->next) {
+		if ((cvar->flags & flags) && cvar->modified) {
+			return true;
+		}
+	}
+	return false;
 }